requirements.txt
# Video processing
moviepy==1.0.3

# Audio processing
librosa==0.9.2
soundfile==0.12.1
pydub==0.25.1

# Machine learning and NLP
torch==1.12.1
torchaudio==0.12.1

transformers==4.44.2 # Hugging Face Transformers
sentencepiece==0.1.99
sacremoses==0.1.1

# File handling and environment variables
python-dotenv==1.0.0

# Utilities
numpy==1.24.3
scipy==1.12.0

# Optional: for better performance with GPU
# torch==2.0.1+cu117
# torchaudio==2.0.2+cu117
